alternative person panel view
