+++ src/main/java/org/ambrogenea/familyview/gui/swing/components/draw/PersonPanel.java
@@ -3,17 +3,18 @@ package org.ambrogenea.familyview.gui.swing.components.draw;
 import java.awt.*;
 import java.awt.event.MouseAdapter;
 import java.awt.image.BufferedImage;
+import java.io.IOException;
 
+import javax.imageio.ImageIO;
 import javax.swing.BorderFactory;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 
-import org.ambrogenea.familyview.domain.DatePlace;
 import org.ambrogenea.familyview.dto.tree.PersonRecord;
+import org.ambrogenea.familyview.enums.Sex;
 import org.ambrogenea.familyview.gui.swing.constant.Fonts;
 import org.ambrogenea.familyview.gui.swing.tools.PersonPanelMouseController;
 import org.ambrogenea.familyview.service.ConfigurationService;
-import org.ambrogenea.familyview.utils.Tools;
 
 /**
  *
@@ -26,15 +27,13 @@ public abstract class PersonPanel extends JPanel {
     protected final PersonRecord person;
     protected final ConfigurationService configuration;
     protected BufferedImage personDiagram;
-    private int fontSize;
+    protected int fontSize;
 
     protected JLabel firstName;
     protected JLabel surName;
     protected JLabel occupation;
-    protected JLabel belowNamesSpace;
     protected JLabel birth;
     protected JLabel birthPlace;
-    protected JLabel belowDatesSpace;
     protected JLabel death;
     protected JLabel deathPlace;
 
@@ -47,16 +46,22 @@ public abstract class PersonPanel extends JPanel {
         initElements();
     }
 
-    protected abstract void loadPictures();
-
-    protected abstract void showPlaces();
-
     protected abstract void addLabels();
+
+    protected abstract void initDateLabels();
+
+    protected abstract void setLabelsOffset();
 
     public void update() {
         this.removeAll();
         initElements();
         revalidate();
+    }
+
+    protected void addEmptyLabel(int yPosition, GridBagConstraints c) {
+        c.ipady = 4;
+        c.gridy = yPosition;
+        add(new JLabel(""), c);
     }
 
     private void initElements() {
@@ -71,20 +76,32 @@ public abstract class PersonPanel extends JPanel {
         addLabels();
     }
 
-    private void initLabels() {
-        belowNamesSpace = new JLabel("", JLabel.CENTER);
-        belowDatesSpace = new JLabel("", JLabel.CENTER);
+    private void loadPictures() {
+        String imagePath;
 
+        if (person.getSex().equals(Sex.MALE)) {
+            imagePath = configuration.getAdultManImagePath();
+        } else {
+            imagePath = configuration.getAdultWomanImagePath();
+        }
+
+        try {
+            personDiagram = ImageIO.read(ClassLoader.getSystemResourceAsStream(imagePath));
+        } catch (IOException e) {
+            System.out.println("Image " + imagePath + " can't be open.");
+        }
+    }
+
+    private void initLabels() {
         occupation = new JLabel("", JLabel.CENTER);
         occupation.setText(person.getOccupation().split(";")[0]);
         occupation.setFont(new Font(Fonts.GENERAL_FONT, Font.PLAIN, fontSize));
-//        occupation.setPreferredSize(new Dimension(configuration.getAdultImageWidth(), fontSize + 2));
 
         initNameLabels();
         initDateLabels();
 
         if (configuration.isShowPlaces() && !birthPlace.getText().isEmpty() || !deathPlace.getText().isEmpty()) {
-            showPlaces();
+            setLabelsOffset();
         }
     }
 
@@ -125,141 +142,6 @@ public abstract class PersonPanel extends JPanel {
         }
     }
 
-    private void initDateLabels() {
-        birth = new JLabel(" ", JLabel.RIGHT);
-        birthPlace = new JLabel("", JLabel.LEFT);
-        death = new JLabel(" ", JLabel.RIGHT);
-        deathPlace = new JLabel("", JLabel.LEFT);
-
-        String birthPlaceString = "";
-        String birthDateString = "";
-        DatePlace birthDatePlace = person.getBirthDatePlace();
-        if (birthDatePlace.getDate() != null) {
-            birthDateString = "\u002A" + birthDatePlace.getLocalizedDate(configuration.getLocale()) + "";
-            if (configuration.isShowPlaces() && !birthDatePlace.getPlace().isEmpty()) {
-                if (configuration.isShortenPlaces()) {
-                    birthPlaceString = "," + SPACE + Tools.cityShortVersion(birthDatePlace.getSimplePlace());
-                } else {
-                    birthPlaceString = "," + SPACE + birthDatePlace.getSimplePlace();
-                }
-            }
-        } else {
-            if (configuration.isShowPlaces() && !birthDatePlace.getPlace().isEmpty()) {
-                if (configuration.isShortenPlaces()) {
-                    birthPlaceString = "\u002A" + Tools.cityShortVersion(birthDatePlace.getSimplePlace());
-                } else {
-                    birthPlaceString = "\u002A" + birthDatePlace.getSimplePlace();
-                }
-                birthPlace.setHorizontalAlignment(JLabel.CENTER);
-            }
-        }
-
-        if (configuration.isShowAge()) {
-            birth.setText(birthDateString);
-            birthPlace.setText(birthPlaceString);
-        } else {
-            birth.setText(birthDateString + birthPlaceString);
-        }
-
-        DatePlace deathDatePlace = person.getDeathDatePlace();
-        if (configuration.isShowAge() && deathDatePlace.getDate() != null) {
-            death.setText("\u2020" + deathDatePlace.getLocalizedDate(configuration.getLocale()) + "");
-            if (configuration.isShowPlaces() && !deathDatePlace.getPlace().isEmpty()) {
-                if (configuration.isShortenPlaces()) {
-                    deathPlace.setText("," + SPACE + Tools.cityShortVersion(deathDatePlace.getSimplePlace()));
-                } else {
-                    deathPlace.setText("," + SPACE + deathDatePlace.getSimplePlace());
-                }
-            }
-        }
-
-        initDateLabelsFont();
-    }
-
-    private void initDateLabelsFont() {
-        birth.setFont(new Font(Fonts.GENERAL_FONT, Font.PLAIN, fontSize));
-        birthPlace.setFont(new Font(Fonts.GENERAL_FONT, Font.PLAIN, fontSize - 1));
-        death.setFont(new Font(Fonts.GENERAL_FONT, Font.PLAIN, fontSize));
-        deathPlace.setFont(new Font(Fonts.GENERAL_FONT, Font.PLAIN, fontSize - 1));
-    }
-
-    protected void addLabels(int verticalShift) {
-        GridBagConstraints c = new GridBagConstraints();
-        c.ipady = 0 - verticalShift;
-        c.weighty = 5;
-        c.gridwidth = 2;
-        add(new JLabel(""), c);
-
-        c.gridy = 1;
-        c.ipady = 0;
-        c.weighty = 0;
-        add(firstName, c);
-        c.gridy = 2;
-        c.ipady = 4;
-        add(surName, c);
-
-        if (configuration.isShowOccupation() && !person.getOccupation().isEmpty()) {
-            c.gridy = 3;
-            c.ipady = 0;
-            add(occupation, c);
-        }
-
-        c.gridy = 4;
-        c.ipady = 4;
-        add(belowNamesSpace, c);
-
-        if (configuration.isShowPlaces()) {
-            c.gridwidth = 1;
-        }
-        c.gridy = 5;
-        c.ipady = 0;
-        c.weighty = 0;
-        if (person.getBirthDatePlace().getDate() != null) {
-            if (!configuration.isShowAge()) {
-                c.gridwidth = 2;
-            }
-            add(birth, c);
-        }
-
-        if (configuration.isShowAge()) {
-            c.gridy = 7;
-            add(death, c);
-        }
-
-        if (configuration.isShowPlaces()) {
-            if (configuration.isShowAge()) {
-                c.gridy = 5;
-                if (person.getBirthDatePlace().getDate() != null) {
-                    c.gridx = 1;
-                } else {
-                    c.gridx = 0;
-                }
-                add(birthPlace, c);
-//                c.gridx = 1;
-                c.gridy = 7;
-                add(deathPlace, c);
-            }
-
-            c.gridx = 0;
-            c.gridwidth = 2;
-        }
-
-        if (configuration.isShowTemple() && !person.isChild() && !person.isLiving()) {
-            c.gridy = 9;
-            c.ipady = 8;
-            add(belowDatesSpace, c);
-            JPanel templeBox = creteTempleBox();
-            c.ipady = 5;
-            c.gridy = 10;
-            add(templeBox, c);
-        }
-
-        c.gridy = 11;
-        c.weighty = 5;
-        c.ipady = 0 + verticalShift;
-        add(new JLabel(""), c);
-    }
-
     @Override
     protected void paintComponent(Graphics g) {
         int imageWidth;
@@ -278,7 +160,7 @@ public abstract class PersonPanel extends JPanel {
         }
     }
 
-    private JPanel creteTempleBox() {
+    protected JPanel creteTempleBox() {
         JPanel templeBox = new JPanel(new GridLayout(1, 3, -1, -1));
         templeBox.setBackground(Color.WHITE);
