+++ src/main/resources/text/CityShortcuts.properties
@@ -25,10 +25,13 @@ horn\u00edradslavice = H. Radslavice
 ostrava-v\u00edtkovice = Ost.-V\u00edtkovice
 ostrava-p\u0159\u00edvoz = Ost.-P\u0159\u00edvoz
 ostrava-kun\u010dice = Ost.-Kun\u010dice
+star\u00e9m\u011bstop.sn\u011b\u017en\u00edkem = St. M\u011bsto p.Sn.
 horn\u00edcerekev = H. Cerekev
 praha-hole\u0161ovice = Hole\u0161ovice
 praha-z\u00e1b\u011bhlice = Z\u00e1b\u011bhlice
 praha-\u0161eberov = \u0160eberov
+praha-jalov\u00fddv\u016fr = Praha-Jal. Dv\u016fr
+moravsk\u00fdp\u00edsek = Mor. P\u00edsek
 
 ol\u0161\u00ednadoslavou = Ol\u0161\u00ed n. Osl.
 obi\u010dtau = Oby\u010dtov
